scraper/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-grpc>=1.20.0
